requirements.txt
# DeepFake detection (Xception-centric). Python 3.11–3.14 tested on CPU; use CUDA wheels for GPU.
torch>=2.0.0
torchvision>=0.15.0
pytorchcv>=0.0.70
albumentations>=2.0.0
opencv-python-headless>=4.9.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
tqdm>=4.65.0
pillow>=10.0.0
numpy>=1.24.0
